# quantize to prevent leaks with float32