--- a/tests/benchmark_svg.nim
+++ b/tests/benchmark_svg.nim
@@ -1,6 +1,8 @@
 import benchy, pixie/fileformats/svg
 
-let data = readFile("tests/fileformats/svg/Ghostscript_Tiger.svg")
+let
+  data = readFile("tests/fileformats/svg/Ghostscript_Tiger.svg")
+  parsed = parseSvg(data)
 
-timeIt "svg parse + render":
-  discard newImage(parseSvg(data))
+timeIt "svg render":
+  discard newImage(parsed)